pyproject.toml
[tool.poetry]
name = "edu_hub"
version = "0.1.0"
description = ""
authors = ["YahyaOmari <yahya.omari@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
django-rest-knox = "^4.1.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
